🐛 Fix wrong reg date
