+++ frontend/src/components/registration-form.tsx
@@ -19,13 +19,14 @@ import {
     Center,
     LinkBox,
     LinkOverlay,
+    Flex,
 } from '@chakra-ui/react';
 import type { SubmitHandler } from 'react-hook-form';
 import { FormProvider, useForm } from 'react-hook-form';
 import { MdOutlineArrowForward } from 'react-icons/md';
 import { useState } from 'react';
 import NextLink from 'next/link';
-import { differenceInHours, format, isBefore, parseISO } from 'date-fns';
+import { differenceInHours, format, isBefore } from 'date-fns';
 import { enUS, nb } from 'date-fns/locale';
 import CountdownButton from '@components/countdown-button';
 import type { Happening, HappeningType, Question } from '@api/happening';
@@ -37,6 +38,7 @@ import useLanguage from '@hooks/use-language';
 import hasOverlap from '@utils/has-overlap';
 import capitalize from '@utils/capitalize';
 import useAuth from '@hooks/use-auth';
+import chooseDate from '@utils/choose-date';
 
 const codeToStatus = (statusCode: number): 'success' | 'warning' | 'error' => {
     if (statusCode === 200) return 'success';
@@ -48,19 +50,6 @@ interface Props {
     happening: Happening;
     type: HappeningType;
 }
-
-const chooseDate = (
-    regDate: string | null,
-    studentGroupRegDate: string | null,
-    userIsEligibleForEarlyReg: boolean,
-): Date => {
-    if (!regDate && !studentGroupRegDate) return new Date();
-    if (!regDate && studentGroupRegDate) return userIsEligibleForEarlyReg ? parseISO(studentGroupRegDate) : new Date();
-    if (regDate && !studentGroupRegDate) return parseISO(regDate);
-    if (regDate && studentGroupRegDate) return parseISO(userIsEligibleForEarlyReg ? studentGroupRegDate : regDate);
-    // Shouldn't be possible to reach this point, but TypeScript doesn't know that
-    return new Date();
-};
 
 const RegistrationForm = ({ happening, type }: Props): JSX.Element => {
     const { isOpen, onOpen, onClose } = useDisclosure();
@@ -149,10 +138,10 @@ const RegistrationForm = ({ happening, type }: Props): JSX.Element => {
         );
 
     return (
-        <Box data-testid="registration-form">
+        <Flex direction="column" gap="5" data-testid="registration-form">
             {!userIsComplete(user) && (
                 <>
-                    <Alert status="warning" borderRadius="0.5rem" mb="5">
+                    <Alert status="warning" borderRadius="0.5rem">
                         <AlertIcon />
                         {isNorwegian
                             ? 'Du må fylle ut all nødvendig informasjon for å kunne melde deg på arrangementer!'
@@ -168,13 +157,13 @@ const RegistrationForm = ({ happening, type }: Props): JSX.Element => {
                 </>
             )}
             {userIsEligibleForEarlyReg && !happening.onlyForStudentGroups && (
-                <Alert status="info" borderRadius="0.5rem" mb="5">
+                <Alert status="info" borderRadius="0.5rem">
                     <AlertIcon />
                     Du kan melde deg på dette arrangementet tidligere enn andre.
                 </Alert>
             )}
             {happening.onlyForStudentGroups && (
-                <Alert status="info" borderRadius="0.5rem" mb="5">
+                <Alert status="info" borderRadius="0.5rem">
                     <AlertIcon />
                     {isNorwegian ? 'Dette er et internt arrangement.' : 'This is a private event.'}
                 </Alert>
@@ -233,7 +222,7 @@ const RegistrationForm = ({ happening, type }: Props): JSX.Element => {
                     </FormProvider>
                 </ModalContent>
             </Modal>
-        </Box>
+        </Flex>
     );
 };
